requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
colorama==0.4.6
comm==0.2.1
coverage==7.4.3
debugpy==1.6.7
decorator==5.1.1
Django==5.0.3
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.1
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
exceptiongroup==1.2.0
executing==2.0.1
factory-boy==3.3.0
Faker==24.0.0
importlib-metadata==7.0.1
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
model-bakery==1.17.0
mysqlclient==2.2.4
nest_asyncio==1.6.0
packaging==23.2
parso==0.8.3
phonenumberslite==8.13.32
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.42
psutil==5.9.0
pure-eval==0.2.2
Pygments==2.17.2
pytest==8.0.2
pytest-cov==4.1.0
pytest-django==4.8.0
pytest-mock==3.12.0
python-dateutil==2.9.0
pytz==2024.1
pywin32==305.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
rpds-py==0.18.0
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tornado==6.3.3
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
wcwidth==0.2.13
wheel==0.41.2
zipp==3.17.0
